requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
discord-interactions==0.4.0
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
protobuf==3.20.3
pycparser==2.21
pydotenv==0.0.7
PyNaCl==1.5.0
python-dotenv==1.0.0
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.2.3
numerize